+++ src/App.scss
@@ -139,6 +139,7 @@ body {
 
     .AccountInfoCell {
       display: table-cell;
+      vertical-align: middle;
 
       button {
         padding: 0.1rem 0.5rem;
