+++ src/config/store.ts
@@ -11,7 +11,7 @@ import BigNumber from "bignumber.js";
 import { reducer as account } from "ducks/account";
 import { reducer as claimAsset } from "ducks/claimAsset";
 import { reducer as claimableBalances } from "ducks/claimableBalances";
-import { reducer as depositAsset } from "ducks/depositAsset";
+import { reducer as sep24DepositAsset } from "ducks/sep24DepositAsset";
 import { reducer as logs } from "ducks/logs";
 import { reducer as sendPayment } from "ducks/sendPayment";
 import { reducer as sendSep31 } from "ducks/sendSep31";
@@ -30,7 +30,7 @@ const reducers = combineReducers({
   account,
   claimAsset,
   claimableBalances,
-  depositAsset,
+  sep24DepositAsset,
   logs,
   sendPayment,
   sendSep31,
