+++ src/methods/sep10Auth/index.ts
@@ -2,4 +2,8 @@ import { start } from "./start";
 import { sign } from "./sign";
 import { send } from "./send";
 
-export { start, sign, send };
+export {
+  start as sep10AuthStart,
+  sign as sep10AuthSign,
+  send as sep10AuthSend,
+};
